collated SVG based icons into new directory
