+++ app/client/components/svgs/arrowIcon.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+export const ArrowIcon = () => (
+  <svg viewBox="0 0 30 30">
+    <path d="M22.8 14.6L15.2 7l-.7.7 5.5 6.6H6v1.5h14l-5.5 6.6.7.7 7.6-7.6v-.9" />
+  </svg>
+);
