+++ app/client/components/svgs/tickIcon.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+
+export const TickIcon = () => (
+  <svg
+    viewBox="0 0 10.79 8.608"
+    css={{ height: "16px", width: "21px", marginRight: "10px" }}
+  >
+    <path d="M2.99 6.58L10.24 0l.55.53-7.8 8.08h-.26L0 4.79l.55-.55 2.44 2.33z" />
+  </svg>
+);
